+++ forge/graphql-client/examples/asset.js
@@ -8,7 +8,6 @@
  */
 
 const GraphqlClient = require('@arcblock/graphql-client');
-const { toAssetAddress } = require('@arcblock/did-util');
 const { fromRandom } = require('@arcblock/forge-wallet');
 
 const endpoint = process.env.FORGE_API_HOST || 'http://127.0.0.1:8210'; // testnet
@@ -22,19 +21,14 @@ const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
     console.log({ owner: owner.toJSON() });
 
     // 1. declare owner
-    let res = await client.sendDeclareTx({
-      tx: {
-        itx: {
-          moniker: 'owner',
-        },
-      },
-      wallet: owner,
-    });
-    console.log('declare.owner.result', res);
+    let hash = await client.declare({ moniker: 'owner', wallet: owner });
+    console.log('declare.owner.result', hash);
     console.log('view owner account', `${endpoint}/node/explorer/accounts/${owner.toAddress()}`);
 
     // 2. create asset for owner
-    const asset = {
+    let assetAddress;
+    // eslint-disable-next-line prefer-const
+    [hash, assetAddress] = await client.createAsset({
       moniker: 'asset',
       readonly: false, // if we want to update the asset, we should set this to false
       transferrable: false,
@@ -45,12 +39,10 @@ const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
           sn: Math.random(),
         },
       },
-    };
-    const assetAddress = toAssetAddress(asset);
-    asset.address = assetAddress;
-    res = await client.sendCreateAssetTx({ tx: { itx: asset }, wallet: owner });
+      wallet: owner,
+    });
     console.log('view asset state', `${endpoint}/node/explorer/assets/${assetAddress}`);
-    console.log('create asset tx', `${endpoint}/node/explorer/txs/${res}`);
+    console.log('create asset tx', `${endpoint}/node/explorer/txs/${hash}`);
 
     // wait for asset state consolidates
     await sleep(5000);
@@ -60,24 +52,20 @@ const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
     console.log('asset state', state);
 
     // 4. update asset
-    res = await client.sendUpdateAssetTx({
-      tx: {
-        itx: {
-          moniker: 'asset_updated',
-          address: assetAddress,
-          data: {
-            typeUrl: 'json',
-            value: {
-              key: 'value updated',
-              sn: Math.random(),
-            },
-          },
+    hash = await client.updateAsset({
+      address: assetAddress,
+      moniker: 'asset_updated',
+      data: {
+        typeUrl: 'json',
+        value: {
+          key: 'value updated',
+          sn: Math.random(),
         },
       },
       wallet: owner,
     });
     console.log('view asset state', `${endpoint}/node/explorer/assets/${assetAddress}`);
-    console.log('update asset tx', `${endpoint}/node/explorer/txs/${res}`);
+    console.log('update asset tx', `${endpoint}/node/explorer/txs/${hash}`);
   } catch (err) {
     console.error(err);
     console.log(JSON.stringify(err.errors));