+++ forge/graphql-client/examples/exchange.js
@@ -18,9 +18,7 @@
  */
 
 const GraphQLClient = require('@arcblock/graphql-client');
-const { toAssetAddress } = require('@arcblock/did-util');
 const { fromRandom } = require('@arcblock/forge-wallet');
-const { fromTokenToUnit } = require('@arcblock/forge-util');
 
 const endpoint = process.env.FORGE_API_HOST || 'http://127.0.0.1:8210'; // testnet
 
@@ -37,37 +35,23 @@ const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
     });
 
     // 1. declare sender
-    let res = await client.sendDeclareTx({
-      tx: {
-        itx: {
-          moniker: 'sender',
-        },
-      },
-      wallet: sender,
-    });
-    console.log('declare.sender.result', res);
+    let hash = await client.declare({ moniker: 'sender', wallet: sender });
+    console.log('declare.sender.result', hash);
     console.log('view sender account', `${endpoint}/node/explorer/accounts/${sender.toAddress()}`);
 
     // 2. declare receiver
-    res = await client.sendDeclareTx({
-      tx: {
-        itx: {
-          moniker: 'receiver',
-        },
-      },
-      wallet: receiver,
-    });
-    console.log('declare.receiver.result', res);
+    hash = await client.declare({ moniker: 'receiver', wallet: receiver });
+    console.log('declare.receiver.result', hash);
     console.log(
       'view receiver account',
       `${endpoint}/node/explorer/accounts/${receiver.toAddress()}`
     );
 
     // 3. create asset for sender
-    const asset = {
+    let assetAddress;
+    // eslint-disable-next-line prefer-const
+    [hash, assetAddress] = await client.createAsset({
       moniker: 'asset',
-      readonly: true,
-      transferrable: true,
       data: {
         typeUrl: 'json',
         value: {
@@ -75,50 +59,36 @@ const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
           sn: Math.random(),
         },
       },
-    };
-    const assetAddress = toAssetAddress(asset);
-    asset.address = assetAddress;
-    res = await client.sendCreateAssetTx({
-      tx: { itx: asset },
       wallet: sender,
     });
-    console.log('create_asset.result', res, assetAddress);
+    console.log('create_asset.result', hash, assetAddress);
     console.log('view asset', `${endpoint}/node/explorer/assets/${assetAddress}`);
-
-    // assemble exchange tx: multisig
-    const exchange = {
-      itx: {
-        to: receiver.toAddress(),
-        sender: {
-          assets: [assetAddress],
-        },
-        receiver: {
-          value: fromTokenToUnit(0),
-        },
-      },
-    };
-
-    console.log('exchange', exchange);
+    await sleep(2000);
 
     // 4.1 Sender: encode and sign the transaction
-    const encoded1 = await client.signExchangeTx({ tx: exchange, wallet: sender });
+    const tx1 = await client.prepareExchange({
+      receiver: receiver.toAddress(),
+      offerAssets: [assetAddress],
+      demandToken: 0,
+      wallet: sender,
+    });
 
     // 4.2 Receiver: do the multi sig
-    const encoded2 = await client.multiSignExchangeTx({
-      tx: Object.assign(encoded1, exchange),
+    const tx2 = await client.finalizeExchange({
+      tx: tx1,
       wallet: receiver,
     });
 
+    console.log({ tx1, tx2 });
+
     // 4.3 Send the exchange tx
-    await sleep(5000);
-    res = await client.sendExchangeTx({
-      tx: encoded2,
+    hash = await client.exchange({
+      tx: tx2,
       wallet: sender,
-      signature: encoded1.signature,
     });
 
-    console.log('exchange hash', res);
-    console.log('exchange tx', `${endpoint}/node/explorer/txs/${res}`);
+    console.log('exchange hash', hash);
+    console.log('exchange tx', `${endpoint}/node/explorer/txs/${hash}`);
   } catch (err) {
     console.error(err);
     console.log(JSON.stringify(err.errors));