+++ forge/graphql-client/examples/migrate_account.js
@@ -6,53 +6,36 @@
  *
  * Run script with: `DEBUG=@arcblock/graphql-client node examples/migrate_account.js`
  */
-const Mcrypto = require('@arcblock/mcrypto');
 const GraphqlClient = require('@arcblock/graphql-client');
-const { fromRandom, WalletType } = require('@arcblock/forge-wallet');
+const { fromRandom } = require('@arcblock/forge-wallet');
 
 const endpoint = process.env.FORGE_API_HOST || 'http://127.0.0.1:8210'; // testnet
 
 const client = new GraphqlClient(`${endpoint}/api`);
 const sleep = timeout => new Promise(resolve => setTimeout(resolve, timeout));
 
-const type = WalletType({
-  role: Mcrypto.types.RoleType.ROLE_ACCOUNT,
-  pk: Mcrypto.types.KeyType.ED25519,
-  hash: Mcrypto.types.HashType.SHA3,
-});
-
 (async () => {
   try {
-    const migrateFrom = fromRandom(type);
-    const migrateTo = fromRandom(type);
+    const migrateFrom = fromRandom();
+    const migrateTo = fromRandom();
 
     // 1. declare migrate_from account
-    let res = await client.sendDeclareTx({
-      tx: {
-        itx: {
-          moniker: `migrate_from_${Math.round(Math.random() * 10000)}`,
-        },
-      },
+    let hash = await client.declare({
+      moniker: `migrate_from_${Math.round(Math.random() * 10000)}`,
       wallet: migrateFrom,
     });
 
     console.log('from account', `${endpoint}/node/explorer/accounts/${migrateFrom.toAddress()}`);
-    console.log('from account tx', `${endpoint}/node/explorer/txs/${res}`);
+    console.log('from account tx', `${endpoint}/node/explorer/txs/${hash}`);
     await sleep(3000);
 
     // 2. migrate: with from's pk/sk
-    res = await client.sendAccountMigrateTx({
-      tx: {
-        itx: {
-          address: migrateTo.toAddress(),
-          pk: migrateTo.publicKey,
-          type,
-        },
-      },
-      wallet: migrateFrom,
+    hash = await client.accountMigrate({
+      from: migrateFrom,
+      to: migrateTo,
     });
 
-    console.log('migrate tx', `${endpoint}/node/explorer/txs/${res}`);
+    console.log('migrate tx', `${endpoint}/node/explorer/txs/${hash}`);
     console.log('to account', `${endpoint}/node/explorer/accounts/${migrateTo.toAddress()}`);
   } catch (err) {
     console.error(err);